backend/app/requirement.txt
fastapi
sqlalchemy
pydentic
python-dotenv
uvicorn